+++ apps/vscode/CHANGELOG.md
@@ -6,6 +6,7 @@
 - Visual Editor: Render pasted markdown text rather than treating it literally
 - Language support (highlighting, completion, etc.) for embedded Matlab code cells
 - Render All command to render and preview all specified formats 
+- Fix crash caused by improper initialization of LSP configuration 
 
 ## 1.86.2 (Release on 26 May 2023)